Separate base template from auth version
