+++ config/application.rb
@@ -15,7 +15,5 @@ insert_into_file 'config/application.rb', before: /^  end/ do
     # (these paths are also autoloaded in development mode)
     # config.eager_load_paths += %W(#{config.root}/lib)
     config.eager_load_paths += %W[#{config.root}/doc]
-
-    config.to_prepare { Devise::Mailer.layout 'mailer' }
   RUBY
 end
