+++ config/template.rb
@@ -1,20 +1,19 @@
 apply 'config/application.rb'
 template 'config/database.example.yml.tt'
 remove_file 'config/database.yml'
-copy_file 'config/routes.rb', force: true
+copy_file "config/routes#{app_type}.rb", 'config/routes.rb', force: true
 copy_file 'config/puma.rb', force: true
 copy_file 'config/sidekiq.yml'
-copy_file 'config/locales/doorkeeper.en.yml'
-
-gsub_file 'config/routes.rb', /  # root 'welcome#index'/ do
-  '  root "home#index"'
-end
 
 copy_file 'config/initializers/generators.rb'
 copy_file 'config/initializers/rotate_log.rb'
 copy_file 'config/initializers/version.rb'
 copy_file 'config/initializers/apipie.rb'
 template 'config/initializers/sidekiq.rb.tt'
+copy_file 'config/initializers/apipie.rb'
+gsub_file 'config/initializers/apipie.rb', /AppName/ do
+  app_name.titleize
+end
 
 gsub_file 'config/initializers/filter_parameter_logging.rb', /\[:password\]/ do
   '%w[password secret session cookie csrf]'
