+++ spec/support/template.rb
@@ -1,2 +1,2 @@
 copy_file 'spec/support/apipie_helper.rb'
-copy_file 'spec/support/request_spec_helper.rb'
+copy_file "spec/support/request_spec_helper#{app_type}.rb", 'spec/support/request_spec_helper.rb'
