+++ template.rb
@@ -9,18 +9,17 @@ def apply_template!
   add_template_repository_to_source_path
 
   template 'Gemfile.tt', force: true
-
   template 'README.md.tt', force: true
+  template 'example.env.tt'
+  template 'ruby-version.tt', '.ruby-version', force: true
+
   remove_file 'README.rdoc'
 
-  template 'example.env.tt'
   copy_file 'editorconfig', '.editorconfig'
   copy_file 'rspec', '.rspec'
   copy_file 'gitignore', '.gitignore', force: true
   copy_file 'overcommit.yml', '.overcommit.yml'
-  template 'ruby-version.tt', '.ruby-version', force: true
   copy_file 'simplecov', '.simplecov'
-
   copy_file 'Procfile'
 
   apply 'Rakefile.rb'
@@ -30,15 +29,15 @@ def apply_template!
   apply 'config/template.rb'
   apply 'doc/template.rb'
   apply 'lib/template.rb'
+  apply 'spec/template.rb'
+  apply 'auth_template.rb' if include_auth_functionality
 
   remove_dir 'test'
-  apply 'spec/template.rb'
 
   git :init unless preexisting_git_repo?
   empty_directory '.git/safe'
 
   run_with_clean_bundler_env 'bin/setup'
-  create_initial_migration
   generate_bundler_binstub
 
   binstubs = %w[
@@ -135,6 +134,17 @@ def staging_hostname
     ask_with_default('Staging hostname?', :blue, 'staging.example.com')
 end
 
+def include_auth_functionality
+  return @include_auth_functionality if defined?(@include_auth_functionality)
+
+  @include_auth_functionality =
+    %w[yes y].include?(ask_with_default('Include user authentiaction? Y/n', :red, 'Y').downcase)
+end
+
+def app_type
+  include_auth_functionality ? '_with_auth' : ''
+end
+
 def gemfile_requirement(name)
   @original_gemfile ||= IO.read('Gemfile')
   req = @original_gemfile[/gem\s+['"]#{name}['"]\s*(,[><~= \t\d.\w'"]*)?.*$/, 1]
@@ -168,26 +178,18 @@ def run_with_clean_bundler_env(cmd)
             else
               run(cmd)
             end
-  unless success
-    puts "Command failed, exiting: #{cmd}"
-    exit(1)
-  end
+  return if success
+
+  puts "Command failed, exiting: #{cmd}"
+  exit(1)
 end
 
 def run_rubocop_autocorrections
-  run_with_clean_bundler_env 'bin/rubocop -a --fail-level A > /dev/null || true'
+  run_with_clean_bundler_env 'bin/rubocop -A --fail-level A > /dev/null || true'
 end
 
-def create_initial_migration
-  return if Dir['db/migrate/**/*.rb'].any?
-
-  now = Time.now.strftime('%Y%m%d%H%M%S').to_i
-  migration_folder = 'db/migrate/'
-
-  copy_file 'db/migrate/create_users.rb', "#{migration_folder}#{now}_create_users.rb"
-  copy_file 'db/migrate/add_devise_to_users.rb', "#{migration_folder}#{now + 1}_add_devise_to_users.rb"
-  copy_file 'db/migrate/add_doorkeeper_to_users.rb', "#{migration_folder}#{now + 2}_add_doorkeeper_to_users.rb"
-  run_with_clean_bundler_env 'bin/rails db:migrate'
+def app_name
+  @app_name ||= @app_path.gsub(/_-/, ' ').titleize
 end
 
 apply_template!
